requirements.txt

mxnet-cu101==1.7.0
gluoncv==0.9.0
matplotlib==3.2.2
opencv-python==4.2.0.34
ImageHash==4.2.0
imutils==0.5.3
dash==1.13.4
dash-bootstrap-components==0.11.0
dash-core-components==1.10.1
dash-deck==0.0.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
plotly==4.9.0
Pillow
numpy>=1.18.3
scikit-image==0.16.2
pandas==1.0.3
gunicorn==20.0.4
dash_bootstrap_components==0.10.7